+++ models/MarketEvent.go
@@ -1,12 +1,8 @@
 package models
 
 import (
-	"errors"
 	"fmt"
-	"io"
-	"net/http"
 	"os"
-	"strings"
 
 	models_pb "github.com/delta/dalal-street-server/proto_build/models"
 	"github.com/delta/dalal-street-server/utils"
@@ -97,40 +93,12 @@ func AddMarketEvent(stockId uint32, headline, text string, isGlobal bool, imageU
 
 	l.Infof("Attempting")
 
-	// Try downloading image first
-	response, err := http.Get(imageURL)
-	if err != nil || response.StatusCode != http.StatusOK {
-		l.Errorf("Error : %v, StatusCode : %d", err, response.StatusCode)
-		if err != nil {
-			return err
-		}
-		return errors.New("NOT OK status code")
-	}
-	defer response.Body.Close()
-
-	// Extract filename
-	var basename = imageURL[strings.LastIndex(imageURL, "/")+1:]
-	var getParamStartIndex = strings.Index(basename, "?")
-	if getParamStartIndex != -1 {
-		basename = basename[:getParamStartIndex]
-	}
-	l.Debugf("ImageURL : %s Basename : %s", imageURL, basename)
-
-	// open file for saving image
-	file, err := os.Create(utils.GetImageBasePath() + basename)
-
+	err := utils.DownloadImage(imageURL)
 	if err != nil {
 		l.Error(err)
 		return err
 	}
-	defer file.Close()
-
-	// copy image to file
-	_, err = io.Copy(file, response.Body)
-	if err != nil {
-		l.Error(err)
-		return err
-	}
+	var basename = utils.GetImageBasename(imageURL)
 
 	db := getDB()
 
@@ -177,40 +145,7 @@ func UpdateMarketEvent(stockId, oldNewsId uint32, headline, text string, isGloba
 			"param_oldNewsId": oldNewsId,
 		})
 
-		// Try downloading image first
-		response, err := http.Get(imageURL)
-		if err != nil || response.StatusCode != http.StatusOK {
-			l.Errorf("Error : %v, StatusCode : %d", err, response.StatusCode)
-			if err != nil {
-				return err
-			}
-			return errors.New("NOT OK status code")
-		}
-		defer response.Body.Close()
-
-		// Extract filename
-		var basename = imageURL[strings.LastIndex(imageURL, "/")+1:]
-		var getParamStartIndex = strings.Index(basename, "?")
-		if getParamStartIndex != -1 {
-			basename = basename[:getParamStartIndex]
-		}
-		l.Debugf("ImageURL : %s Basename : %s", imageURL, basename)
-
-		// open file for saving image
-		file, err := os.Create(utils.GetImageBasePath() + basename)
-
-		if err != nil {
-			l.Error(err)
-			return err
-		}
-		defer file.Close()
-
-		// copy image to file
-		_, err = io.Copy(file, response.Body)
-		if err != nil {
-			l.Error(err)
-			return err
-		}
+		var basename = utils.GetImageBasename(imageURL)
 
 		me := &MarketEvent{
 			StockId:   stockId,
@@ -225,12 +160,29 @@ func UpdateMarketEvent(stockId, oldNewsId uint32, headline, text string, isGloba
 		l.Infof("Attempting to update existing market event ")
 
 		OldEvent := &MarketEvent{}
-		if err = db.First(OldEvent, oldNewsId).Error; err != nil {
+		if err := db.First(OldEvent, oldNewsId).Error; err != nil {
 			l.Error(err)
 			return err
 		}
 
-		if err = db.Model(&OldEvent).Updates(me).Error; err != nil {
+		// If Image has to be changed
+		if OldEvent.ImagePath != basename {
+			// Delete old Image
+			err := os.Remove(utils.GetImageBasePath() + OldEvent.ImagePath)
+			if err != nil {
+				l.Error(err)
+				return err
+			}
+
+			// Download new image
+			err = utils.DownloadImage(imageURL)
+			if err != nil {
+				l.Error(err)
+				return err
+			}
+		}
+
+		if err := db.Model(&OldEvent).Updates(me).Error; err != nil {
 			l.Error(err)
 			return err
 		}
